feat(web/common): update component import
